+++ guillotina_cms/api/components.py
@@ -1,7 +1,10 @@
 from guillotina import configure
 from guillotina.interfaces import IAbsoluteURL
 from guillotina.interfaces import IResource
-from guillotina.interfaces import IContainer
+from guillotina.component import query_utility
+from guillotina.utils import get_content_path
+from guillotina.utils import get_content_depth
+from guillotina.interfaces import ICatalogUtility
 from guillotina.interfaces import IDatabase
 from guillotina.api.service import Service
 from guillotina_cms.interfaces import ICMSLayer
@@ -52,6 +55,13 @@ class Breadcrumbs(Service):
         }
 
 
+def recursive_fill(mother_list, pending_dict):
+    for element in mother_list:
+        if element['@id'] in pending_dict:
+            element['items'] = pending_dict[element['@id']]
+            recursive_fill(element['items'], pending_dict)
+
+
 @configure.service(
     context=IResource, method='GET',
     permission='guillotina.AccessContent', name='@navigation',
@@ -80,17 +90,55 @@ class Breadcrumbs(Service):
 class Navigation(Service):
 
     async def __call__(self):
-        result = []
-        container = self.request.container
-        async for content in container.async_values():
-            if IResource.providedBy(content):
-                result.append({
-                    'title': content.title,
-                    '@id': IAbsoluteURL(content, self.request)()
-                })
+        search = query_utility(ICatalogUtility)
+        path = get_content_path(self.context)
+        depth = get_content_depth(self.context) + 1
+        max_depth = None
+        if 'expand.navigation.depth' in self.request.rel_url.query:
+            max_depth = str(int(self.request.rel_url.query['expand.navigation.depth']) + depth)
+            musts = [
+                {'range': {'depth': {'gte': depth}}},
+                {'range': {'depth': {'lte': max_depth}}}]
+        else:
+            musts = [{'term': {'depth': depth}}]
+        query = {
+            '_source': {
+                'includes': ['title', '_id', 'depth', 'parent_uuid', '@absolute_url', 'type_name', 'path'],
+            },
+            'query': {
+                'bool': {
+                    'must': musts,
+                }
+            },
+            'sort': [{'position_in_parent': 'desc'}],
+        }
+        call_params = {
+            'container': self.request.container,
+            'path': path,
+            'query': query,
+            'size': 100
+        }
+        result = await search.get_by_path(**call_params)
+
+        pending_dict = {}
+        for brain in result['member']:
+            brain_serialization = {
+                'title': brain['title'],
+                '@id': brain['@absolute_url']
+            }
+            pending_dict.setdefault(brain['parent_uuid'], []).append(brain_serialization)
+
+        parent_uuid = self.context.uuid
+        if parent_uuid not in pending_dict:
+            final_list = []
+        else:
+            final_list = pending_dict[parent_uuid]
+        if max_depth is not None:
+            recursive_fill(final_list, pending_dict)
+
         return {
             "@id": self.request.url.human_repr(),
-            "items": result
+            "items": final_list
         }